style: input with icon - dir 구조로 변경
