+++ src/components/common/Input/InputWithIcon/index.ts
@@ -1,0 +1,1 @@
+export * from "./InputWithIcon";
